change get post to get in an orderly manner, upload images to cloudinary, like post, change @Data to -getter -setter to solve ToString issue

+++ src/main/java/api/socialPlatform/ApiForSocialApp/dto/PostResponseDto.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 public class PostResponseDto {
     private List<String> postImageUrls;
     private String postDescription;
-    private int likes;
+    private long likes;
     private Date createdAt;
     private Date updatedAt;
     private UUID userId;
